move abc9_ops -reintegrate into its own pass
